fix: separate subview to profileview
