--- IntroVerse/Presentation/Nick/InfoView.swift
@@ -1,5 +1,0 @@
-import UIKit
-
-final class InfoView: UIView {
-    
-}
